+++ src/client/app/components/UI/NavBar.jsx
@@ -1,4 +1,3 @@
-/* eslint-disable react/forbid-component-props */
 import React from 'react';
 import { NavLink } from 'react-router-dom';
 // eslint-disable-next-line no-unused-vars
@@ -11,51 +10,12 @@ import NavProfile from './navProfile';
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import showElement from '../../../../server/utils/console/showElement';
 import UserBalance from './userBalance';
+import NavList from './navlist';
 
 export default function NavBar() {
-  const isLogged = useSelector(getLoginStatus());
-
   return (
     <div className="navbar">
-      <nav>
-        <ul className='nav-list'>
-          {isLogged ?
-            <>
-              <li className='nav-item'>
-                <NavLink
-                  aria-current="page"
-                  className={({ isActive }) => (isActive ? 'nav-link navbar-brand' : 'nav-link')}
-                  to="/analytics"
-                >
-                  Аналитика
-                </NavLink>
-              </li>
-              <li className='nav-item'>
-                <NavLink
-                  aria-current="page"
-                  className={({ isActive }) => (isActive ? 'nav-link navbar-brand' : 'nav-link')} //TODO убрать выделение, посмотреть другие варианты.
-                  to="/operations"
-                >
-                  Операции
-                </NavLink>
-              </li>
-            </> :
-            <>
-              <li className='nav-item'>
-                <NavLink
-                  aria-current="page"
-                  className={({ isActive }) => (isActive ? 'nav-link navbar-brand' : 'nav-link')}
-                  to="/"
-                >
-                  Главная
-                </NavLink>
-              </li>
-              <li className='nav-item login'>
-                <NavLink className='nav-link' to="/login" >Войти</NavLink>
-              </li>
-            </>}
-        </ul>
-      </nav>
+      <div className="navlist--top"><NavList /></div>
       <div className="user-panel">
         <UserBalance />
         <NavProfile />
