+++ src/client/app/styles/navbar.css
@@ -6,7 +6,7 @@ nav {
   background-color: rgba(255, 255, 255, 0.3);
   padding: .5rem 1.5rem;
 
-  display: grid;
+  display: flex;
   grid-auto-flow: column;
   /* justify-content: space-between;
   align-items: center;
@@ -14,8 +14,13 @@ nav {
   width: 100%; */
 }
 
+.navlist--bottom {
+  display: none;
+}
+
 .nav-list {
   display: flex;
+  justify-content: space-around;
   align-items: center;
   gap: 1rem;
   list-style: none;
@@ -40,8 +45,12 @@ nav {
 }
 
 @media (max-width: 600px) {
-  nav {
+  .navlist--top {
     display: none;
+  }
+  .navlist--bottom {
+    display: flex;
+    justify-content: center;
   }
   .user-panel {
     min-width: 100%;
